Add GetLatestCart feature and improve code readability (UPDATED DATABASE)
Introduced a new method `GetLatestCart` in `ICartService` and implemented it in `CartService` to retrieve the latest cart for a given account. Added a new endpoint `CheckOut/{accountId}` in `CartController` to utilize this method. Enhanced `EpicGameDbContext` with JSON conversion for `Permission` property in `Role` entity and updated `OnModelCreating` for better readability. Refactored `CartRepository` to include related entities in `GetAllCartFromAccountId` method. Removed unnecessary constructors and reformatted code for improved readability.